CORE-6331 Created commons theme
Moved the Iplant and internal anchor appearances into the new themes
module.
