+++ applications/belphegor/src/main/java/org/iplantc/admin/belphegor/belphegor.gwt.xml
@@ -22,7 +22,7 @@
     <set-property name="gwt.logging.popupHandler" value="DISABLED"/>
     <set-property name="gwt.logging.consoleHandler" value="DISABLED"/>
 
-    <replace-with class="org.iplantc.de.commons.client.appearance.widgets.IPlantAnchorDefaultAppearance">
+    <replace-with class="org.iplantc.de.theme.base.client.commons.widgets.IPlantAnchorDefaultAppearance">
         <when-type-is class="org.iplantc.de.commons.client.widgets.IPlantAnchorAppearance" />
     </replace-with>
