+++ modules/notifications/src/main/java/org/iplantc/de/systemMessages/SystemMessages.gwt.xml
@@ -3,5 +3,6 @@
 <module>
 
     <inherits name='com.google.gwt.user.User'/>
+    <inherits name="org.iplantc.de.commons.ui-commons" />
 
 </module>
