--- modules/ui-commons-module/src/main/java/org/iplantc/de/commons/client/appearance/widgets/package-info.java
@@ -1,4 +1,0 @@
-/**
- * This package holds the widget appearance implementations.
- */
-package org.iplantc.de.commons.client.appearance.widgets;
