--- modules/ui-commons-module/src/main/java/org/iplantc/de/commons/client/events/CollaboratorsLoadedEventHandler.java
@@ -1,16 +1,0 @@
-package org.iplantc.de.commons.client.events;
-
-import com.google.gwt.event.shared.EventHandler;
-
-/**
- * 
- * Defines handler for Collaborators laoded event
- * 
- * @author sriram
- * 
- */
-public interface CollaboratorsLoadedEventHandler extends EventHandler {
-
-    public void onLoad(CollaboratorsLoadedEvent event);
-
-}
