+++ modules/ui-commons-module/src/main/java/org/iplantc/de/commons/client/presenter/Presenter.java
@@ -5,3 +5,4 @@ import com.google.gwt.user.client.ui.HasOneWidget;
 public abstract interface Presenter {
     public abstract void go(final HasOneWidget container);
 }
+
