+++ modules/ui-commons-module/src/main/java/org/iplantc/de/commons/ui-commons.gwt.xml
@@ -1,8 +1,3 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  When updating your version of GWT, you should also update this DTD reference,
-  so that your app can take advantage of the latest GWT module capabilities.
--->
 <!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit 2.5.0//EN"
   "http://google-web-toolkit.googlecode.com/svn/tags/2.5.0/distro-source/core/src/gwt-module.dtd">
 <module>
@@ -20,12 +15,6 @@
     <inherits name='org.iplantc.de.resources.ui-resources' />
     <inherits name='org.iplantc.de.DiscoveryEnvironmentCommon' />
 
-    <!-- Default Appearances -->
-    <replace-with class="org.iplantc.de.commons.client.appearance.widgets.InternalAnchorDefaultAppearance">
-        <when-type-is class="org.iplantc.de.commons.client.widgets.InternalAnchor.Appearance"/>
-    </replace-with>
-    
-    
     <source path="client"/>
 
 </module>
